Create new component Spendings; Add more stalls to the list
