+++ src/common/Spendings/index.module.css
@@ -1,0 +1,7 @@
+
+.buttonWrapper{
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  margin-bottom: 30px;
+}
