+++ src/common/index.jsx
@@ -10,6 +10,7 @@ export { default as MenuItems } from './MenuItems'
 export { default as Navbar } from './Navbar'
 export { default as PersonCard } from './PersonCard'
 export { default as ScrollDown } from './ScrollDown'
+export { default as Spendings } from './Spendings'
 export { default as Sponsors } from './Sponsors'
 export { default as Tile } from './Tile'
 export { default as UpButton } from './UpButton'
