+++ src/pages/data.js
@@ -216,4 +216,21 @@ export const stalls = [
   {stallName: 'Hogs Heaven', type: 'food',},
   {stallName: 'The Plant Kitchen', type: 'food',},
   {stallName: 'Burger Bus', type: 'food',},
+  {stallName: 'Little Rainbow Jewels ', type: 'commercial',},
+  {stallName: 'Glo Emporium ', type: 'commercial',},
+  {stallName: 'Sent with Pride', type: 'commercial',},
+  {stallName: 'Lynne Powell- beaded goods', type: 'commercial',},
+  {stallName: 'Sweets n Treats', type: 'commercial',},
+  {stallName: 'Fabling Boutique', type: 'commercial',},
+  {stallName: 'Paintings by Lucia', type: 'commercial',},
+  {stallName: 'LGBT Llama', type: 'commercial',},
+  {stallName: 'Gilly Jean Crafts', type: 'commercial',},
+  {stallName: 'Hazel Gibbs', type: 'commercial',},
+  {stallName: 'Cara Designs', type: 'commercial',},
+  {stallName: 'Gonks', type: 'commercial',},
+  {stallName: 'Concrete prairie - merch', type: 'commercial',},
+  {stallName: 'Pride Merch', type: 'commercial',},
+  {stallName: 'Pride Silent Auction ', type: 'commercial',},
+  {stallName: 'Pride Face Painting', type: 'commercial',},
+  {stallName: 'HCRG Care Group', type: 'wellbeing',},
 ]
